[refactor] 도메인 레이어 패키지 구조 구분
- aniaml 패키지에서 Animal, Medical, Memo를 각각의 패키지로 구분하고,
관련된 유스케이스를 패키지별로 분리함.
