+++ mobile/app/src/main/java/com/lacuc/pets/domain/animal/memo/AddMemoUseCase.kt
@@ -1,4 +1,4 @@
-package com.lacuc.pets.domain.animal
+package com.lacuc.pets.domain.animal.memo
 
 import com.lacuc.pets.data.animal.AnimalRepository
 import com.lacuc.pets.data.animal.Memo
