+++ mobile/app/src/main/java/com/lacuc/pets/domain/animal/memo/GetMemoUseCase.kt
@@ -1,7 +1,7 @@
-package com.lacuc.pets.domain.animal
+package com.lacuc.pets.domain.animal.memo
 
 import com.lacuc.pets.data.animal.AnimalRepository
-import com.lacuc.pets.ui.animal.detail.AnimalDetailMemoItem
+import com.lacuc.pets.domain.animal.AnimalDetailMemoItem
 import javax.inject.Inject
 
 class GetMemoUseCase @Inject constructor(private val repository: AnimalRepository) {
