+++ mobile/app/src/main/java/com/lacuc/pets/ui/animal/add/AddAnimalViewModel.kt
@@ -3,7 +3,7 @@ package com.lacuc.pets.ui.animal.add
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.lacuc.pets.data.animal.Animal
-import com.lacuc.pets.domain.animal.AddAnimalUseCase
+import com.lacuc.pets.domain.animal.animal.AddAnimalUseCase
 import com.lacuc.pets.util.SingleLiveEvent
 import com.lacuc.pets.util.safeValue
 import javax.inject.Inject
