+++ mobile/app/src/main/java/com/lacuc/pets/ui/animal/detail/AnimalDetailAddMemoViewModel.kt
@@ -3,7 +3,7 @@ package com.lacuc.pets.ui.animal.detail
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.lacuc.pets.data.animal.Memo
-import com.lacuc.pets.domain.animal.AddMemoUseCase
+import com.lacuc.pets.domain.animal.memo.AddMemoUseCase
 import com.lacuc.pets.util.SingleLiveEvent
 import com.lacuc.pets.util.safeValue
 import javax.inject.Inject
