+++ mobile/app/src/main/res/layout/item_animal_detail_detail.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="item"
-            type="com.lacuc.pets.ui.animal.detail.AnimalDetailDetailItem" />
+            type="com.lacuc.pets.domain.animal.AnimalDetailDetailItem" />
     </data>
 
     <LinearLayout
