+++ mobile/app/src/main/res/layout/item_animal_detail_medical.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="item"
-            type="com.lacuc.pets.ui.animal.detail.AnimalDetailMedicalItem" />
+            type="com.lacuc.pets.domain.animal.AnimalDetailMedicalItem" />
     </data>
 
     <LinearLayout
